% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_topology.R
\name{compute_weights_n_orders}
\alias{compute_weights_n_orders}
\title{Compute weights for N orders (internal function)}
\usage{
compute_weights_n_orders(
  neighbor_orders,
  sigma,
  center_weight,
  weights_input,
  avg_dist,
  adaptive_sigma_factor
)
}
\description{
Compute weights for N orders (internal function)
}
\keyword{internal}
