% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_topology.R
\name{estimate_avg_neighbor_distance}
\alias{estimate_avg_neighbor_distance}
\title{Estimate average distance between first-order neighbours (internal)}
\usage{
estimate_avg_neighbor_distance(grid_proj, first_neighbors, sample_size)
}
\description{
Estimate average distance between first-order neighbours (internal)
}
\keyword{internal}
