% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grid.R
\name{get_utm_crs}
\alias{get_utm_crs}
\title{Get an appropriate UTM CRS for a study area}
\usage{
get_utm_crs(study_area)
}
\arguments{
\item{study_area}{sf object representing the study area. May be in any CRS;
it will be transformed to WGS84 (EPSG:4326) internally for the centroid
calculation.}
}
\value{
Character string with the UTM CRS, e.g. `"EPSG:32630"` for UTM 30N.
}
\description{
Determines the appropriate UTM coordinate reference system (CRS) for a given
study area based on the longitude/latitude of its centroid.
}
\examples{
\dontrun{
library(sf)
study_area <- st_sf(geometry = st_sfc(
  st_polygon(list(matrix(c(-5, 35, 5, 35, 5, 45, -5, 45, -5, 35),
                         ncol = 2, byrow = TRUE))),
  crs = 4326
))
get_utm_crs(study_area)  # "EPSG:32630"
}
}
