% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_wrapper.R
\name{process_district_all_vars_n_orders_wrapper}
\alias{process_district_all_vars_n_orders_wrapper}
\title{C++ wrapper for N-order smoothing (internal)}
\usage{
process_district_all_vars_n_orders_wrapper(
  variable_values,
  neighbors,
  weights,
  hex_indices,
  var_names
)
}
\description{
C++ wrapper for N-order smoothing (internal)
}
\keyword{internal}
