% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_wrapper.R
\name{validate_smoothing_inputs}
\alias{validate_smoothing_inputs}
\title{Validate inputs to the smoothing functions (internal)}
\usage{
validate_smoothing_inputs(variable_values, neighbors, weights, var_names)
}
\description{
Validate inputs to the smoothing functions (internal)
}
\keyword{internal}
