% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_topology.R
\name{compute_neighbors_n_orders}
\alias{compute_neighbors_n_orders}
\title{Compute neighbors for N orders (internal function)}
\usage{
compute_neighbors_n_orders(grid_proj, neighbor_orders)
}
\description{
Compute neighbors for N orders (internal function)
}
\keyword{internal}
