% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grid.R
\name{create_single_grid}
\alias{create_single_grid}
\title{Actually make the grid for one area}
\usage{
create_single_grid(
  study_area,
  grid_size,
  type,
  projection_crs,
  return_crs,
  area_id = NULL,
  check_size = TRUE,
  max_cells = 1e+06
)
}
\description{
Actually make the grid for one area
}
\keyword{internal}
